Added a common header file that keeps track of the Flow algorithm.

+++ Links/Buffer.h
@@ -10,7 +10,7 @@ private:
     std::queue<Packet*> buffer; 
 public:
     int capacityUsed;
-    unsigned int nextFree;
+    double nextFree;
     void push(Packet *p);
     void pop();
     Packet *front();
